requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anthropic>=0.40.0
anyio==4.12.0
blinker==1.8.2
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==46.0.3
distro==1.9.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.12.0
Mako==1.3.5
MarkupSafe==2.1.5
openai==2.14.0
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow>=10.0.0
psycopg2-binary==2.9.6
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.2.0
pytesseract==0.3.13
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.34
stripe==5.5.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.2
Werkzeug==2.3.6
opencv-python-headless
pytesseract
